Add class S3Client. Move utils to utils folder
Refactoring
