+++ config.py
@@ -1,2 +1,9 @@
+""" Global parameters """
+
+S3_BUCKET = 'kaggle-mining-process'
+MODEL_PATH = 'models/'
+MODEL_FILE = 'model.bin'
+SCALER_FILE = 'scaler.b'
 DATA_YEAR = 2017
 EXPERIMENT_NAME = "my-experiment-6"
+LOG_LEVEL = 'TRACE'  # TODO describe others
