+++ train.py
@@ -1,30 +1,28 @@
-import io
-import os
-import glob
-import pickle
+''' Train and register best model '''
 
-import boto3
-import numpy as np
+import os
+import pickle
+from pathlib import Path
+
 import mlflow
-import pandas as pd
 from dotenv import load_dotenv
 from loguru import logger
 from xgboost import XGBRegressor
 from sklearn.tree import DecisionTreeRegressor
-from sklearn.metrics import (
-    r2_score,
-    max_error,
-    mean_squared_error,
-    mean_absolute_error,
-    mean_absolute_percentage_error,
-)
 from sklearn.ensemble import RandomForestRegressor
 from sklearn.linear_model import Ridge, ElasticNet, LinearRegression
 from sklearn.preprocessing import MinMaxScaler
 from sklearn.model_selection import GridSearchCV
 
 import config
-from logger import Logger
+from model import (
+    eval_metrics,
+    get_test_data,
+    get_train_data,
+    validate_month,
+    preprocess_data,
+)
+from utils.logger import Logger
 
 N_FOLDS = 5
 N_JOBS = 8
@@ -35,36 +33,6 @@ PREPROCCESS_TYPE = 'minmax_scaler no3rd'
 
 TRAIN_DATA_MONTH = 7
 TEST_DATA_MONTH = 8
-
-# Do not use '% Iron Concentrate' because of highly correlated(0.8) with target
-# and do not use datetime
-AGG_FEATURES = [
-    'Starch Flow',
-    'Amina Flow',
-    'Ore Pulp Flow',
-    'Ore Pulp pH',
-    'Ore Pulp Density',
-    'Flotation Column 01 Air Flow',
-    'Flotation Column 02 Air Flow',
-    'Flotation Column 03 Air Flow',
-    'Flotation Column 04 Air Flow',
-    'Flotation Column 05 Air Flow',
-    'Flotation Column 06 Air Flow',
-    'Flotation Column 07 Air Flow',
-    'Flotation Column 01 Level',
-    'Flotation Column 02 Level',
-    'Flotation Column 03 Level',
-    'Flotation Column 04 Level',
-    'Flotation Column 05 Level',
-    'Flotation Column 06 Level',
-    'Flotation Column 07 Level',
-]
-LAB_FEATURES = ['% Iron Feed', '% Silica Feed']
-AGG_FEATURES = [clm.replace(' ', '_').lower() for clm in AGG_FEATURES]
-LAB_FEATURES = [clm.replace(' ', '_').lower() for clm in LAB_FEATURES]
-TARGET = '%_silica_concentrate'
-
-Logger()
 
 
 models = {
@@ -92,7 +60,7 @@ models = {
     },
     'RandomForestRegressor': {
         'model': RandomForestRegressor(
-            criterion='mae', n_jobs=N_JOBS, random_state=SEED
+            criterion='absolute_error', n_jobs=N_JOBS, random_state=SEED
         ),
         'params': {
             'max_depth': (3, 4, 5),
@@ -106,7 +74,7 @@ models = {
             'eval_metric': ["mae"],
             'n_estimators': [200, 500, 1000],
             'max_depth': [3, 5, 7],
-            'alpha': [0.1, 1, 5, 10, 100],
+            'alpha': [0.1, 1, 5, 10, 50, 75, 100, 150, 200],
         },
     },
 }
@@ -120,107 +88,11 @@ def init_mlflow():
     logger.info(f"{mlflow.list_experiments()}")
 
 
-def init_s3_client():
-    session = boto3.session.Session()
-    s3 = session.client(
-        service_name='s3',
-        endpoint_url='https://storage.yandexcloud.net',
-        region_name=os.environ['AWS_REGION'],
-        aws_access_key_id=os.environ['AWS_ACCESS_KEY_ID'],
-        aws_secret_access_key=os.environ['AWS_SECRET_ACCESS_KEY'],
-    )
-    return s3
-
-
-def get_list_data(source='local'):
-    if source == 'local':
-        files = glob.glob('*.parquet')
-    elif source == 's3':
-        s3 = init_s3_client()
-        files = [
-            key['Key']
-            for key in s3.list_objects(Bucket='kaggle-mining-process')['Contents']
-        ]
-    return files
-
-
-def pd_read_s3_parquet(file, bucket, s3_client=None) -> pd.DataFrame:
-    '''Read single parquet file from S3'''
-    obj = s3_client.get_object(Bucket=bucket, Key=file)
-    return pd.read_parquet(io.BytesIO(obj['Body'].read()))
-
-
-def validate_month(month) -> str:
-    if not isinstance(month, str):
-        month = str(month)
-    if int(month) < 10:
-        month = '0' + month
-    if int(month) > 12:
-        pass  # TODO
-    return month
-
-
-def get_train_data(month, year=str(config.DATA_YEAR)) -> pd.DataFrame:
-    s3_client = init_s3_client()  # TODO
-    all_data = get_list_data(source='s3')
-    filenames = [
-        file for file in all_data if file <= f"mining_data_{year}-{month}.parquet"
-    ]
-    df = pd.concat(
-        [
-            pd_read_s3_parquet(file, 'kaggle-mining-process', s3_client=s3_client)
-            for file in filenames
-        ]
-    )
-    return df, filenames
-
-
-def get_test_data(month, year='2017') -> pd.DataFrame:
-    s3_client = init_s3_client()  # TODO
-    filename = f"mining_data_{year}-{month}.parquet"
-    df = pd_read_s3_parquet(filename, 'kaggle-mining-process', s3_client=s3_client)
-    return df, filename
-
-
-def preprocess_data(df: pd.DataFrame) -> pd.DataFrame:
-    df.columns = [clm.replace(' ', '_').lower() for clm in df.columns]
-    create_feature_dict = {
-        **{clm: ['last'] for clm in LAB_FEATURES},
-        **{clm: [np.mean, np.max, np.min] for clm in AGG_FEATURES},
-    }
-    X = (
-        df.groupby(pd.Grouper(key='datetime', axis=0, freq='H'))
-        .agg(create_feature_dict)
-        .reset_index()
-        .dropna()
-    )
-    X.columns = ['_'.join(col) for col in X.columns]
-    y = (
-        df[['datetime', TARGET]]
-        .groupby(pd.Grouper(key='datetime', axis=0, freq='H'))
-        .agg('last')
-        .reset_index()
-        .dropna()
-    )
-
-    return X.drop('datetime_', axis=1), y.drop('datetime', axis=1)
-
-
-def eval_metrics(actual, pred):
-    rmse = np.sqrt(mean_squared_error(actual, pred))
-    mae = mean_absolute_error(actual, pred)
-    mape = mean_absolute_percentage_error(actual, pred)
-    r2 = r2_score(actual, pred)
-    max_er = max_error(actual, pred)
-    metrics = {'rmse': rmse, 'mae': mae, 'mape': mape, 'r2': r2, 'max_error': max_er}
-    return metrics
-
-
 if __name__ == '__main__':
-    load_dotenv()  # load environment variables from .env
+    Logger(filename='mlops.log', level=config.LOG_LEVEL)
+    load_dotenv(override=True)  # load environment variables from .env
 
     init_mlflow()
-
     month = validate_month(TRAIN_DATA_MONTH)
     df_train, train_filenames = get_train_data(month=month)
     logger.info(f"Train: {df_train.shape}")
@@ -228,9 +100,7 @@ if __name__ == '__main__':
     logger.info(f"Train after preprocess: {X_train.shape}{y_train.shape}")
     scaler = MinMaxScaler()
     X_train = scaler.fit_transform(X_train)
-    with open("models/scaler.b", "wb") as f:
-        pickle.dump(scaler, f)
-    # mlflow.log_artifact("scaler.b", artifact_path="scaler")
+    Path(f"{config.MODEL_PATH}{config.MODEL_FILE}").write_bytes(pickle.dumps(scaler))
 
     month = validate_month(TEST_DATA_MONTH)
     df_test, test_filename = get_test_data(month=month)
@@ -259,13 +129,14 @@ if __name__ == '__main__':
             clf.fit(X_train, y_train)
             model = clf.best_estimator_
             cv_score = -clf.best_score_
-            with open('models/model.bin', 'wb') as f_out:
-                pickle.dump(model, f_out)
             test_metrics = eval_metrics(y_test, model.predict(X_test))
             logger.info(clf.best_params_)
             logger.info(f"cross-val {SCORING}: {cv_score}")
             for metric, value in test_metrics.items():
                 logger.info(f"{metric}: {value}")
+            Path(f"{config.MODEL_PATH}{config.MODEL_FILE}").write_bytes(
+                pickle.dumps(model)
+            )
             mlflow.set_tag("model", model_name)
             mlflow.log_param("model", model_name)
             mlflow.log_param("preprocess", PREPROCCESS_TYPE)
@@ -277,7 +148,8 @@ if __name__ == '__main__':
             mlflow.log_param("test-data-file", test_filename)
             mlflow.log_metric(f"cv_{SCORING}", cv_score)
             mlflow.log_metrics(test_metrics)
-            # mlflow.log_artifact(local_path="models/model.bin", artifact_path="models_pickle")
+            # mlflow.log_artifact(local_path="models/model.bin",
+            #  artifact_path="models_pickle")
             if m == 'XGBoost':
                 mlflow.xgboost.log_model(model, artifact_path="models_mlflow")
             else:
