+++ utils/logger.py
@@ -10,27 +10,20 @@ import sys
 
 from loguru import logger
 
-
-class Singleton(type):
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
+from utils.singleton import Singleton
 
 
 class Logger(metaclass=Singleton):
     created = None
 
-    def __init__(self, filename='log.log'):
+    def __init__(self, filename='log.log', level='TRACE'):
         if not self.created:
             logger.remove()
-            logger.add(sys.stderr, format="{time} | {level} | {message}")
+            logger.add(sys.stderr, format="{time} | {level} | {message}", level=level)
             logger.add(
                 f"{filename}",
                 format="{time} | {level} | {message}",
-                level='TRACE',
+                level=level,
                 rotation="100 MB",
             )
             self.created = True
